+++ install.ts
@@ -1,17 +1,30 @@
-import { brightGreen, gray, red } from "./deps/colors.ts";
-import { checkDenoVersion, getImportMap } from "./ci.ts";
+import { brightGreen, cyan, gray, green, red } from "./deps/colors.ts";
+import { checkDenoVersion, getImportMap } from "./core/utils.ts";
+import { encode } from "./deps/base64.ts";
 
-checkDenoVersion();
+const denoInfo = checkDenoVersion();
 
+if (denoInfo) {
+  console.log("----------------------------------------");
+  console.error(red("Error installing Lume"));
+  console.log(`Lume needs Deno ${green(denoInfo.minimum)} or greater`);
+  console.log(`Your current version is ${red(denoInfo.current)}`);
+  console.log(`Run ${cyan(denoInfo.command)} and try again`);
+  console.log("----------------------------------------");
+  Deno.exit(1);
+}
+
+const importMap = `data:application/json;base64,${
+  encode(JSON.stringify(getImportMap()))
+}`;
 const cli = new URL("./cli.ts", import.meta.url).href;
-
 const process = Deno.run({
   cmd: [
     Deno.execPath(),
     "install",
     "--unstable",
     "-Af",
-    `--import-map=${await getImportMap()}`,
+    `--import-map=${importMap}`,
     `--no-check`,
     "--name=lume",
     cli,